MoveData writing, research, and experiments
- Currently,  the move and repoint version doesn't work very well. Sometime evething thinks it is at 0pp. read something about PP data being stored around the movedata somewhere?
- Editing moves in place works perfectly though!
- Currently, when expanding movesets,  type definitions, and moving the move data, things seem to crash
- The ??? type appears to be unaffected by move power, and seems to not be a special type OR a physical type.
- Move effects fully researched and documented
- May have to re-add ??? as a fully new type. Wish I could find the phys / special definition...

+++ PokemonRandomizer/PokemonRandomizer/Backend/DataStructures/MoveData.cs
@@ -20,7 +20,7 @@ namespace PokemonRandomizer.Backend.DataStructures
 
         public enum MoveEffect
         {
-            //Need to do science here
+            //See Move effects.txt
         }
 
         public enum Targets
@@ -39,15 +39,15 @@ namespace PokemonRandomizer.Backend.DataStructures
 
         public Move move;
         public MoveEffect effect;
-        public int power;
+        public byte power;
         public PokemonType type;
-        public int accuracy;
-        public int pp;
-        public int effectChance;
+        public byte accuracy;
+        public byte pp;
+        public byte effectChance;
         public Targets targets;
         // This byte is signed, If it is strictly less than 0x80 (128) then it is positive. 
         // If not, the actual value equals: -1 * (256 - Current Value). Example: value 0xFE (254) must be treated as -2 instead.
-        public int priority;
+        public byte priority;
         // Flags documentation
         public bool Contact { get => flags[0]; }               // 0 - This moves makes contact with the target.
         public bool AffectedByProtect { get => flags[1]; }     // 1 - This move is affected by Protect.
@@ -69,7 +69,7 @@ namespace PokemonRandomizer.Backend.DataStructures
             targets = (Targets)rom.ReadByte();
             priority = rom.ReadByte();
             flags = new BitArray(new byte[]{ rom.ReadByte() });
-            rom.Skip(3);
+            rom.Skip(3); // three bytes of 0x00
         }
 
         public override string ToString()
